install/requirements_cuda.txt
tqdm
numpy
matplotlib
einops
torch
cupy-cuda12x